requirements.txt
milvus-haystack
farm-haystack[inference, crawler]
streamlit
markdown
python-dotenv